Add error logging for folders
If a folder does not exist, Ansible outputs an error stating that Nullvalue does not have a function MoveInto. This is due to folder not existing and the value returned from pyvmomi being null.

This will inform the user if the folder name/path does not exist so they can troubleshoot accordingly.